fix(proxy): consolidate rust control-plane auth route
