--- litellm/proxy/rust_control_plane/__init__.py
@@ -1,1 +1,0 @@
-"""Internal routes exposed by the Python control plane for Rust data planes."""
