--- tests/test_litellm/proxy/rust_control_plane/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for Rust control-plane endpoints."""
